refactor: modularize DataFrameEntry
Decided to give DataFrameEntry it's own file
